예외처리

+++ library-app/src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -33,6 +33,7 @@ public class UserController {
        * query 함수의 첫번째 인자는 해당 sql 을, 두번째 인자는 rowMapper 이다.
        * GetMapping 이니까 가져오는거고, query 를 통해 테이블에서 가져온다.
        * resultSet 으로 각 맞는 타입을 받아서 new UserResponse 로 반환한다.
+       * jdbcTemplate.query 구문은 LIST 를 반환한다.
        * */
         return jdbcTemplate.query(sql, (resultSet, rowNum) -> {
             long id = resultSet.getLong("id");
@@ -45,25 +46,42 @@ public class UserController {
 
    @PatchMapping("/user")
    public void updateUser(@RequestBody UserUpdateRequest dto) {
+        String readSql = "SELECT * FROM user WHERE id = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql , (re, rowNum) -> 0 , dto.getId()).isEmpty();
+        // (re, rowNum) -> 0 은 결과가 있다면, 0을 반환한다. 그렇지 않으면 .isEmpty() 를 타서
+       // isUserNotExist 가 true 가 된다.
+
+        if (isUserNotExist) {
+            throw new IllegalArgumentException();
+        }
+
         String sql = "UPDATE user SET name = ? WHERE id = ?";
         jdbcTemplate.update(sql, dto.getName() , dto.getId());
    }
 
    @DeleteMapping("/user")
-    public void deleteUser(@RequestParam String name) {
-        /*
-        * @RequestParam 는 파라미터 {{baseurl}}/user?name=jw
-        * */
-        String sql = "DELETE FROM user WHERE name = ?";
-        jdbcTemplate.update(sql , name);
+   public void deleteUser(@RequestParam String name) {
+       String readSql = "SELECT * FROM user WHERE name = ?";
+       boolean isUserNotExist = jdbcTemplate.query(readSql , (re, rowNum) -> 0 , name).isEmpty();
+
+       if (isUserNotExist) {
+           throw new IllegalArgumentException();
+       }
+
+        //@RequestParam 는 파라미터 {{baseurl}}/user?name=jw
+       String sql = "DELETE FROM user WHERE name = ?";
+       jdbcTemplate.update(sql , name);
    }
 
     @DeleteMapping("/user/{id}")
     public void deleteUserById(@PathVariable Long id) {
-        /*
-         * @PathVariable 는 주소 옆에다가.
-         * */
+        //@PathVariable 는 주소 옆에다가.
         String sql = "DELETE FROM user WHERE id = ?";
         jdbcTemplate.update(sql , id);
+    }
+
+    @GetMapping("/user/error-test")
+    public void errorTest() {
+        throw new IllegalArgumentException();
     }
 }
